refactor: Merge BottomButtonsStripe component files
